always initialize foreground/background with QRgb colors
More fallout from dbf0ca2b: TextStyle.foreground and background are of type
QRgb so we can't initialize them directly to enum values; so now we need to
access the ColorPalette instance to get the default colors.
This bad initialization caused a bug: with very rapid output from the
shell (such as when running find), some lines of text would have foreground
color 8, nearly black, and fail to be changed again later, for some reason.

Also removed stuff that isn't in use: the intense color table, and
the xtermColor accessor.