Have a crack at fixing command keys
I'm not sure how this worked before (if it did), but it certainly
doesn't now. I get a key event for 'Key_C', with modifier meta (on mac,
ctrl on Linux) set -- the return here means it would have always been
swallowed which is definitely not we want.

Fixes #2 (without any fallout, I hope).