Cursor: Fix thinko in moveToLine
This manifested as completely broken output in e.g. top, which makes
extensive use of VPA to draw multiple lines of output, which ended up
wrapping (soft) rather than rendering multiple lines.

I'd really like to see some tests for Parser -> Cursor interaction, but
that's going to be quite a job to do.