screen: Move font setting C++ side
For some crazy reason, Qt on Linux decides it's going to reset the pixel size
by one *after* we've done initial setup. This change doesn't propagate
to all the existing UI, and regardless of that, seems a little crazy.

Let's address this by moving it C++ side, and ensuring that *we* keep
full control of when the properties of it change.

This means we have no solution for high DPI screens right now, but let's
cross that bridge when it comes up.