screen: Try move to a single source of truth for fonts
The C++ approach (in 6ce0e5fea7d082a5e5c217235f4af2865ed7ca25, which we
are reverting) was a bandaid for a more annoying problem in that we
weren't reacting to font changes properly. Try to fix that, and at the
same time, move all measurements to operate on font metrics rather than
a bit of a mix (MonoText already used them, the QML side didn't).

There are still some weird problems with certain font pixel sizes, but
15 seems to work pretty well for me so far, fingers crossed.

This improves some cursor positioning bugs for me on Linux.