Block: Fix dangerous API
Returning a ref to a QString is extremely dangerous, and should not be
done outside of API that does so in an obvious sort of manner (e.g.
QStringRef-returning midRef type functions).

I doubt this has any obvious problems from a quick grep, but neither
does the code in question (text selection) seem critical enough to
warrant such treatment in the first place.