Rewrite screen size handling a little. Contributes to #10.
We revert 443ce6b67049b55317d34e74e1c138f84e594d16, as I'm pretty sure
with additional testing that this was misguided at best.

New approach here: we make size props asynchronous, and for now, tie
them in with the usual event dispatch mechanism. Future TODO: tie
TerminalScreen's updatePolish to update size if need be.

On top of that, combine YatPty's width and height resize into a single
operation, and send it early while we're processing our own resize.

There's some gremlins lurking here still, no doubt, but this seems to be
a step in the right direction.

This reverts commit 443ce6b67049b55317d34e74e1c138f84e594d16.